requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
langchain-community
pinecone
python-dotenv
pandas